Rename News categorization assets
