+++ assets/ar/news_categorization/SANADAlArabiya_GPT4_FewShot.py
@@ -1,6 +1,6 @@
 import random
 
-from llmebench.datasets import NewsCatAkhbaronaDataset
+from llmebench.datasets import SANADAlArabiyaDataset
 from llmebench.models import OpenAIModel
 from llmebench.tasks import NewsCategorizationTask
 
@@ -10,7 +10,7 @@ random.seed(1333)
 
 def config():
     return {
-        "dataset": NewsCatAkhbaronaDataset,
+        "dataset": SANADAlArabiyaDataset,
         "dataset_args": {},
         "task": NewsCategorizationTask,
         "task_args": {},
@@ -28,9 +28,9 @@ def config():
             "max_tries": 30,
         },
         "general_args": {
-            "data_path": "data/news_categorization/SANAD_akhbarona_news_cat_test.tsv",
+            "data_path": "data/news_categorization/SANAD_alarabiya_news_cat_test.tsv",
             "fewshot": {
-                "train_data_path": "data/news_categorization/SANAD_akhbarona_news_cat_train.tsv"
+                "train_data_path": "data/news_categorization/SANAD_alarabiya_news_cat_train.tsv"
             },
         },
     }
@@ -75,21 +75,23 @@ def prompt(input_sample, examples):
 def post_process(response):
     label = response["choices"][0]["message"]["content"]
 
+    label_list = config()["model_args"]["class_labels"]
     label_fixed = label.lower()
     label_fixed = label_fixed.replace("category: ", "")
     label_fixed = label_fixed.replace("science/physics", "tech")
     label_fixed = label_fixed.replace("health/nutrition", "medical")
-    label_fixed = label_fixed.replace("nutrition", "medical")
-    label_fixed = label_fixed.replace("health", "medical")
 
-    if len(label_fixed.split("\s+")) > 1:
-        label_fixed = label_fixed.split("\s+")[0]
-    label_fixed = random.choice(label_fixed.split("/")).strip()
-    if "science/physics" in label_fixed:
+    if label_fixed in label_list:
+        label_fixed = label_fixed
+
+    elif "science/physics" in label_fixed:
         label_fixed = label_fixed.replace("science/physics", "tech")
-    if label_fixed.startswith("culture"):
+    elif label_fixed.startswith("culture"):
         label_fixed = label_fixed.split("(")[0]
-
         label_fixed = label_fixed.replace("culture.", "culture")
+    elif "/" in label:
+        label_fixed = random.choice(label_fixed.split("/")).strip()
+    else:
+        label_fixed = None
 
     return label_fixed
